requirements.txt
Flask_SQLAlchemy==2.4.4
Flask_SocketIO==5.0.1
Flask_WTF==0.14.3
Flask_Cors==3.0.10
Flask_Login==0.5.0
Flask_Bcrypt==0.7.1
SQLAlchemy==1.3.23
Flask_JWT==0.3.2
Flask_DebugToolbar==0.11.0
Werkzeug==1.0.1
aiohttp==3.7.4.post0
Flask==1.1.2
WTForms==2.3.3
PyJWT==2.1.0

